Kindly inform our friendly daemons not to mess with external servers

+++ daemon-bash/src/main/java/io/minecloud/daemon/MineCloudDaemon.java
@@ -31,6 +31,7 @@ import io.minecloud.models.nodes.NodeRepository;
 import io.minecloud.models.server.Server;
 import io.minecloud.models.server.ServerMetadata;
 import io.minecloud.models.server.ServerRepository;
+import io.minecloud.models.server.type.ServerLaunchType;
 import io.minecloud.models.server.type.ServerType;
 import org.mongodb.morphia.query.Query;
 import redis.clients.jedis.Jedis;
@@ -95,6 +96,11 @@ public class MineCloudDaemon {
                         return;
 
                     Server server = mongo.repositoryBy(Server.class).findFirst(stream.readString());
+                    
+                    //Sanity: Ignore external servers
+                    if (server.type().launchType() == ServerLaunchType.EXTERNAL) {
+                        return;
+                    }
 
                     if (!server.node().name().equals(node)) {
                         MineCloud.logger().log(Level.SEVERE, "Invalid request was sent to kill a server " +
@@ -170,7 +176,12 @@ public class MineCloudDaemon {
 
                     MessageInputStream stream = message.contents();
                     Server server = mongo.repositoryBy(Server.class).findFirst(stream.readString());
-
+                    
+                    //Ignore external servers
+                    if (server.type().launchType() == ServerLaunchType.EXTERNAL) {
+                        return;
+                    }
+                    
                     if (!server.node().name().equals(node))
                         return;
 
@@ -183,7 +194,14 @@ public class MineCloudDaemon {
                         return;
 
                     MessageInputStream stream = message.contents();
-                    File file = new File("/var/minecloud/" + stream.readString());
+                    Server server = mongo.repositoryBy(Server.class).findFirst(stream.readString());
+                    
+                    //Ignore external servers
+                    if (server.type().launchType() == ServerLaunchType.EXTERNAL) {
+                        return;
+                    }
+                    
+                    File file = new File("/var/minecloud/" + server.entityId());
 
                     if (file.exists()) {
                         file.delete();
@@ -209,6 +227,10 @@ public class MineCloudDaemon {
                     .collect(Collectors.toList());
 
             nodeServers.stream().filter((s) -> s.ramUsage() != -1).forEach((server) -> {
+                //Sanity: Ignore external servers
+                if (server.type().launchType() == ServerLaunchType.EXTERNAL) {
+                    return;
+                }
                 File runDir = new File("/var/minecloud/" + server.name());
 
                 if (!runDir.exists()) {
@@ -242,6 +264,9 @@ public class MineCloudDaemon {
 
             try (Jedis jedis = this.redis.grabResource()) {
                 nodeServers.stream().filter((s) -> s.ramUsage() != -1).forEach(server ->  {
+                    if (server.type().launchType() == ServerLaunchType.EXTERNAL) {
+                        return;
+                    }
                     Map<String, String> hResult = jedis.hgetAll("server:" + server.entityId());
 
                     if (hResult == null || hResult.isEmpty()) {
